refactor: define `receiver` module
Receiver will be used by transfer exchanger.
