+++ kernel/src/device/pci/xhci/exchanger/mod.rs
@@ -1,3 +1,4 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 pub mod command;
+pub mod receiver;
